@@ -1,5 +1,5 @@
 {
-  "name": "project",
+  "name": "ineed",
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
@@ -8,54 +8,74 @@
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "precache": "sw-precache",
+    "precache-build": "sw-precache --verbose --config=sw-precache-config.js",
+    "upload": "ng build --aot --prod && gulp css js && npm run precache-build && gulp js && firebase deploy --only hosting"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.3.2",
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/core": "^4.3.1",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/router": "^4.3.1",
-    "angular2-image-upload": "^0.6.6",
-    "angular2-uuid": "^1.1.1",
-    "angularfire2": "^4.0.0-rc.1",
+    "@angular/animations": "^4.4.4",
+    "@angular/cdk": "2.0.0-beta.12",
+    "@angular/common": "^4.4.4",
+    "@angular/compiler": "^4.4.4",
+    "@angular/core": "^4.4.4",
+    "@angular/flex-layout": "^2.0.0-beta.9",
+    "@angular/forms": "^4.4.4",
+    "@angular/http": "^4.4.4",
+    "@angular/material": "2.0.0-beta.12",
+    "@angular/platform-browser": "^4.4.4",
+    "@angular/platform-browser-dynamic": "^4.4.4",
+    "@angular/router": "^4.4.4",
+    "@covalent/core": "^1.0.0-beta.8-1",
+    "angular2-image-upload": "^1.0.0-rc.0",
+    "angular2-jwt": "^0.2.3",
+    "angular2-text-mask": "^8.0.4",
+    "angularfire2": "4.0.0-rc.1",
+    "auth0-js": "^8.10.1",
+    "auth0-lock": "^10.23.1",
     "bootstrap": "^3.3.7",
-    "core-js": "^2.4.1",
-    "firebase": "^4.1.5",
-    "fuse.js": "^3.0.5",
+    "core-js": "^2.5.1",
+    "firebase": "4.1.5",
+    "font-awesome": "^4.7.0",
+    "fuse.js": "^3.2.0",
+    "hammerjs": "^2.0.8",
     "jquery": "^3.2.1",
-    "ngx-bootstrap": "^1.8.0",
-    "ngx-modialog": "^3.0.2",
+    "md2": "0.0.29",
+    "ngx-bootstrap": "^1.9.3",
+    "ngx-modialog": "^3.0.4",
     "ngx-pagination": "^3.0.1",
-    "primeng": "^4.1.1",
+    "primeng": "^4.2.2",
     "primeui": "^4.1.15",
-    "rxjs": "^5.1.0",
-    "zone.js": "^0.8.4"
+    "rxjs": "^5.4.3",
+    "sendbird": "^3.0.42",
+    "zone.js": "^0.8.18"
   },
   "devDependencies": {
-    "@angular/cli": "1.2.0",
-    "@angular/compiler-cli": "^4.0.0",
-    "@angular/language-service": "^4.0.0",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~3.0.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
+    "@angular/cli": "1.4.6",
+    "@angular/compiler-cli": "^4.4.4",
+    "@angular/language-service": "^4.4.4",
+    "@types/jasmine": "~2.6.0",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.0.34",
+    "codelyzer": "~3.2.1",
+    "gulp": "^3.9.1",
+    "gulp-concat": "^2.6.1",
+    "gulp-csso": "^3.0.0",
+    "gulp-gzip": "^1.4.0",
+    "gulp-uglify": "^3.0.0",
+    "jasmine-core": "~2.8.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~1.7.1",
+    "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-coverage-istanbul-reporter": "^1.3.0",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.1.2",
-    "ts-node": "^3.0.6",
-    "tslint": "^5.3.2",
-    "typescript": "^2.4.2"
+    "sw-precache": "^5.2.0",
+    "ts-node": "~3.3.0",
+    "tslint": "~5.7.0",
+    "typescript": "~2.5.3"
   }
 }
