@@ -1,7 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { environment } from '../environments/environment';
-import { FormsModule} from '@angular/forms';
+import { FormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
 
 import { AngularFireModule } from 'angularfire2';
@@ -12,11 +12,20 @@ import { AppComponent } from './app.component';
 import { HomeComponent } from './home/home.component';
 import { AuthService } from './shared/services/services-auth/auth.service';
 import { VerifyAuth } from './shared/services/verify-auth/verify-auth';
+import { NoopAnimationsModule } from '@angular/platform-browser/animations';
+import { CallbackComponent } from './callback/callback.component';
+import { Auth0Service } from './shared/services/auth0-service/auth0.service';
+import { MatToolbarModule } from '@angular/material';
+import { MatButtonModule } from '@angular/material';
+import { NoConflictStyleCompatibilityMode, CompatibilityModule } from '@angular/material';
+import { MatMenuModule } from '@angular/material';
+import { MatIconModule } from '@angular/material';
 
-@NgModule({ 
+@NgModule({
   declarations: [
     AppComponent,
-    HomeComponent
+    HomeComponent,
+    CallbackComponent
   ],
   imports: [
     BrowserModule,
@@ -25,9 +34,17 @@ import { VerifyAuth } from './shared/services/verify-auth/verify-auth';
     HttpModule,
     AngularFireModule.initializeApp(environment.firebase),
     AngularFireDatabaseModule,
-    AngularFireAuthModule
+    AngularFireAuthModule,
+    NoopAnimationsModule,
+    MatToolbarModule,
+    MatButtonModule,
+    NoConflictStyleCompatibilityMode,
+    CompatibilityModule,
+    MatMenuModule,
+    MatIconModule
   ],
-  providers: [AuthService,VerifyAuth],
+  providers: [AuthService, VerifyAuth, Auth0Service],
   bootstrap: [AppComponent]
 })
-export class AppModule { }
+export class AppModule {
+}
