@@ -1,4 +1,4 @@
-import { TestBed, async, inject } from '@angular/core/testing';
+import { TestBed, inject } from '@angular/core/testing';
 
 import { MustBeLoggedInGuard } from './must-be-logged-in.guard';
 
