@@ -3,12 +3,13 @@ import { Routes, RouterModule } from '@angular/router';
 import { ChatComponent } from './chat.component';
 
 const routes: Routes = [
-	{ path: '', component: ChatComponent },
- 	{ path: '**', redirectTo: '', pathMatch: 'full'}
+  {path: '', component: ChatComponent},
+  {path: '**', redirectTo: '', pathMatch: 'full'}
 ];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule]
 })
-export class ChatRoutingModule { }
+export class ChatRoutingModule {
+}
