@@ -1,4 +1,5 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnInit} from '@angular/core';
+import {ProductsService} from '../../admin/products-dashboard/services/products.service';
 
 @Component({
   selector: 'app-main-dashboard',
@@ -6,10 +7,17 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./main-dashboard.component.css']
 })
 export class MainDashboardComponent implements OnInit {
+  // Doughnut
+  public doughnutChartLabels: string[] = ['Total de produtos', 'Total de Lojas', 'Total de funcionarios'];
+  public doughnutChartData: number[] = [1000, 5, 5];
+  public doughnutChartType: string = 'doughnut';
 
-  constructor() { }
+  constructor() {
+
+  }
 
   ngOnInit() {
   }
 
 }
+
