@@ -6,20 +6,21 @@ import { MustBeAdminGuard } from './guards/must-be-admin/must-be-admin.guard';
 import { MustBeLoggedInGuard } from '../shared/guards/must-be-logged-in/must-be-logged-in.guard';
 
 const routes: Routes = [
-  { 
-  	path: 'dashboard',
+  {
+    path: 'dashboard',
     component: ShopkeeperComponent,
     canActivate: [MustBeLoggedInGuard],
-    children : [
+    children: [
       { path: '', redirectTo: 'home', pathMatch: 'full' },
       { path: 'home', component: MainDashboardComponent },
-      { path: 'admin', canActivate: [MustBeAdminGuard], children: [
+      {
+        path: 'admin', canActivate: [MustBeAdminGuard], children: [
           { path: 'products', loadChildren: './admin/products-dashboard/products-dashboard.module#ProductsDashboardModule' },
           { path: 'employees', loadChildren: './admin/employees-dashboard/employees-dashboard.module#EmployeesDashboardModule' },
           { path: 'stores', loadChildren: './admin/stores-dashboard/stores-dashboard.module#StoresDashboardModule' }
-        ] 
+        ]
       }
-    ] 
+    ]
   },
   { path: 'chat', canActivate: [MustBeLoggedInGuard], loadChildren: './chat/chat.module#ChatModule' },
   { path: '**', redirectTo: 'dashboard', pathMatch: 'full' }
