@@ -7,7 +7,8 @@ import { Component, OnInit } from '@angular/core';
 })
 export class MessageEmailVerificationComponent implements OnInit {
 
-  constructor() { }
+  constructor() {
+  }
 
   ngOnInit() {
   }
