@@ -9,16 +9,17 @@ import { IsLoggedAuthGuard } from './guards/is-logged-auth/is-logged-auth.guard'
 import { SubscribeComponent } from './subscribe.component';
 
 const routes: Routes = [
-	{ path: '', component: SubscribeComponent, children: [
-	    { path: 'signup', component: SignUpComponent, canActivate: [IsLoggedAuthGuard] },
-	    { path: 'signin', component: SignInComponent, canActivate: [IsLoggedAuthGuard] },
-	    { path: 'emailverication', component: MessageEmailVerificationComponent, canActivate: [IsLoggedAuthGuard] },
-	    { path: 'adminverication', component: MessageAdminVerificationComponent, canActivate: [IsLoggedAuthGuard] },
-	    { path: 'forgot', component: PasswordForgotComponent, canActivate: [IsLoggedAuthGuard] },
-	    { path: '', redirectTo: 'signin', pathMatch: 'full' },
-	    { path: '**', redirectTo: 'signin', pathMatch: 'full' }
-  	]
-	}
+  {
+    path: '', component: SubscribeComponent, children: [
+    {path: 'signup', component: SignUpComponent, canActivate: [IsLoggedAuthGuard]},
+    {path: 'signin', component: SignInComponent, canActivate: [IsLoggedAuthGuard]},
+    {path: 'verification/email', component: MessageEmailVerificationComponent, canActivate: [IsLoggedAuthGuard]},
+    {path: 'verification/admin', component: MessageAdminVerificationComponent, canActivate: [IsLoggedAuthGuard]},
+    {path: 'forgot', component: PasswordForgotComponent, canActivate: [IsLoggedAuthGuard]},
+    {path: '', redirectTo: 'signin', pathMatch: 'full'},
+    {path: '**', redirectTo: 'signin', pathMatch: 'full'}
+  ]
+  }
 ];
 
 @NgModule({
@@ -26,4 +27,5 @@ const routes: Routes = [
   exports: [RouterModule],
   providers: [IsLoggedAuthGuard]
 })
-export class SubscribeRoutingModule { }
+export class SubscribeRoutingModule {
+}
