@@ -7,10 +7,8 @@ import { Component, OnInit } from '@angular/core';
 })
 export class SubscribeComponent implements OnInit {
 
-  // Começa no metodo de sign Up
-  private method: string = 'signUp';
-
-  constructor() { }
+  constructor() {
+  }
 
   ngOnInit() {
   }
