@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
 
 import { SubscribeRoutingModule } from './subscribe-routing.module';
@@ -10,14 +10,20 @@ import { PasswordForgotComponent } from './password-forgot/password-forgot.compo
 import { MessageEmailVerificationComponent } from './message-email-verification/message-email-verification.component';
 import { MessageAdminVerificationComponent } from './message-admin-verification/message-admin-verification.component';
 import { SubscribeComponent } from './subscribe.component';
+import { FlexLayoutModule } from '@angular/flex-layout';
+import { MatIconModule, MatListModule, MatInputModule, MatCardModule, MatButtonModule, NoConflictStyleCompatibilityMode, CompatibilityModule } from '@angular/material';
 
 @NgModule({
   imports: [
     CommonModule,
     FormsModule,
     HttpModule,
-    SubscribeRoutingModule
+    ReactiveFormsModule,
+    SubscribeRoutingModule,
+    FlexLayoutModule,
+    MatIconModule, MatListModule, MatCardModule, MatButtonModule, NoConflictStyleCompatibilityMode, CompatibilityModule, MatInputModule
   ],
   declarations: [SignUpComponent, SignInComponent, PasswordForgotComponent, MessageEmailVerificationComponent, MessageAdminVerificationComponent, SubscribeComponent]
 })
-export class SubscribeModule { }
+export class SubscribeModule {
+}
